el perfil para el comunero tiene más datos
