+++ app/src/main/java/com/example/pueblosdb/ProfileComuFragment.java
@@ -2,6 +2,7 @@ package com.example.pueblosdb;
 
 import static android.content.Context.MODE_PRIVATE;
 
+import android.app.AlertDialog;
 import android.content.SharedPreferences;
 import android.graphics.Color;
 import android.graphics.drawable.ColorDrawable;
@@ -10,7 +11,7 @@ import android.os.Bundle;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
-import android.app.AlertDialog;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -25,16 +26,18 @@ import com.google.firebase.auth.EmailAuthProvider;
 
 import java.util.HashSet;
 
-public class UserProfileFragment extends Fragment {
+
+public class ProfileComuFragment extends Fragment {
     private SharedPreferences prefs;
 
-    public UserProfileFragment() {
+    public ProfileComuFragment() {
         // Required empty public constructor
     }
 
     @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_user_profile, container, false);
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_profile_comu, container, false);
     }
 
     @Override
@@ -42,17 +45,17 @@ public class UserProfileFragment extends Fragment {
         super.onViewCreated(view, savedInstanceState);
 
         prefs = requireActivity().getSharedPreferences(getString(R.string.prefs_file), MODE_PRIVATE);
-        TextView tv1 = view.findViewById(R.id.user_emailvisualizer);
+        TextView tv1 = view.findViewById(R.id.emailvisualizer);
         tv1.setText(prefs.getString("email", "No hay datos"));
-        TextView tv2 = view.findViewById(R.id.user_namevisualizer);
+        TextView tv2 = view.findViewById(R.id.namevisualizer);
         tv2.setText(prefs.getString("name", "No hay datos"));
-        TextView tv3 = view.findViewById(R.id.user_surnamevisualizer);
+        TextView tv3 = view.findViewById(R.id.surnamevisualizer);
         tv3.setText(prefs.getString("surname", "No hay datos"));
-        TextView tv4 = view.findViewById(R.id.user_showCargo);
+        TextView tv4 = view.findViewById(R.id.showCargo);
         tv4.setText(prefs.getString("cargo", "No hay datos"));
 
 
-        Button changeEmail = view.findViewById(R.id.user_changeEmail);
+        Button changeEmail = view.findViewById(R.id.changeEmail);
         changeEmail.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -75,7 +78,7 @@ public class UserProfileFragment extends Fragment {
             }
         });
 
-        Button changePassword = view.findViewById(R.id.user_changePassword);
+        Button changePassword = view.findViewById(R.id.changePassword);
         changePassword.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -98,7 +101,7 @@ public class UserProfileFragment extends Fragment {
             }
         });
 
-        Button saveChanges = view.findViewById(R.id.user_saveChanges);
+        Button saveChanges = view.findViewById(R.id.saveChanges);
         saveChanges.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -106,7 +109,7 @@ public class UserProfileFragment extends Fragment {
             }
         });
 
-        Button deleteAccount = view.findViewById(R.id.user_deleteAccount);
+        Button deleteAccount = view.findViewById(R.id.deleteAccount);
         deleteAccount.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -138,4 +141,6 @@ public class UserProfileFragment extends Fragment {
             }
         });
     }
+
+
 }
